Fix MSVC narrowing errors exposed by /W4 /WX
With --verbose the Windows failure is finally legible:

  platform/src/windows/local_ipc_win.cpp(64,47): error C2220: the following
  warning is treated as an error
    warning C4267: conversion from 'size_t' to 'uint32_t', possible loss of data

This is fallout from making meminfo_set_strict_warnings MSVC-aware. MSVC had
been silently ignoring -Werror, so these narrowing conversions were never
fatal; /W4 /WX makes them errors. Correcting the flags was right, and these
are real truncations, so they are fixed rather than re-suppressed.

Earlier notes localised the failure to the FlatBuffers custom rule in
common/. That was wrong: "Building Custom Rule" was simply the last line
MSBuild emitted before the parallel build aborted. The error was always in
meminfo_platform, and only became visible once MSBuild printed its summary.

local_ipc_win.cpp gets the same hardening the POSIX backend already received
in this branch: message sizes bounded by kMaxIpcMessageBytes (an unvalidated
length prefix could otherwise drive a multi-gigabyte allocation), lengths
narrowed explicitly after validation, write failures checked, and the
per-connection handling extracted into serve_client().

MSBuild stops at the first failing target, so the remaining targets had never
been compiled by MSVC at all. Rather than discover them one CI round-trip at
a time, the tree was scanned for the same defect class and four more sites
fixed:

- uv_buf_init() takes an unsigned int length; announcer.cpp, client_session.cpp,
  gpu_session.cpp and gpu_client.cpp all passed size_t (memclient was already
  fixed earlier in this branch).
- uv_strerror() takes int; client_session.cpp and gpu_session.cpp passed ssize_t.
- discovery_daemon.cpp stored int config values into uint16_t ports; now
  narrowed through a helper that rejects out-of-range values instead of
  truncating them.

Verifying this needed care. spdlog's bundled fmt sets
`#pragma GCC diagnostic ignored "-Wconversion"` and never pops it, which
silently disables the analysis for every translation unit that includes
spdlog -- so an initial "clean" scan was meaningless. The scan was redone
against a scratch copy of spdlog with that suppression neutralised, and
validated by confirming it reproduces exactly the four lines CI reported
(64, 67, 88, 91) on the pre-fix file. Windows-only code was type-checked
through a Win32 shim that models DWORD as 32-bit so the narrowing matches
MSVC. Linux build and all 45 tests still pass.

+++ discoveryd/src/announcer.cpp
@@ -135,7 +135,7 @@ void Announcer::on_timer(uv_timer_t* handle) {
     SendReqCtx* ctx = new SendReqCtx;
     ctx->buf_base = data;
     
-    uv_buf_t buf = uv_buf_init(ctx->buf_base, size);
+    uv_buf_t buf = uv_buf_init(ctx->buf_base, static_cast<unsigned int>(size));
     ctx->req.data = ctx;
     
     uv_udp_send(&ctx->req, &self->udp_handle_, &buf, 1,
